Referee updates.
Currently implementing within the Referee Class the functionality to print off each of the matches with Team Numbers so that final scores and winners can be assigned to the matches. Running into a problem where the Referee Class is assigning Team Objects to play more than once. Work time: 8 hours,  40 mins.